Fixed some more advanced warnings
- missing parentheses
- not matching data types
- initialization order

+++ vm/src/interpreter/interpreter_primitives.cpp
@@ -792,7 +792,7 @@ void Squeak_Interpreter::lookup_in_obsoleteNamedPrimitiveTable(Oop functionName,
 fn_t Squeak_Interpreter::munge_arguments_and_load_function_from_plugin_on_main(Oop functionName, Object_p& fno, int functionLength,
                                                                                Oop  moduleName, Object_p& mno, int moduleLength) { 
   char fn[10000], mod[10000];
-  assert_always(functionLength < sizeof(fn)  &&  moduleLength < sizeof(mod));
+  assert_always((size_t)functionLength < sizeof(fn)  &&  (size_t)moduleLength < sizeof(mod));
   strncpy( fn, fno->as_char_p() + Object::BaseHeaderSize, functionLength);  fn[functionLength] = '\0';
   strncpy(mod, mno->as_char_p() + Object::BaseHeaderSize,   moduleLength); mod[  moduleLength] = '\0';
   fn_t addr = The_Interactions.load_function_from_plugin(Logical_Core::main_rank, fn, mod);
